requirements.txt
psycopg==3.2.12
python-decouple==3.8
typing_extensions==4.15.0
